Started changing modals to 'react-native-modal', because it supports much better animations. First instance is when clicking on an image on the tour list place interface.
Reverted changes to tourPlaceList, because this is handled in another branch.

Changed TouchableHighlight, which turns black when clicked, to TouchableOpacity.

Merged TourPlaceList style changes with this branch.

Small change to sync to master branch

changed background color in ImageViewer to be consistent with other headers

Added margin to down arrow in TourRecord so that it isn't so close to the edge. Also added a little fade effect when clicking on a tour list place.

Small changes to sync with master.