#102 Crash on iOS when viewing marker callout
The app was crashing when trying to view a callout for a map marker 
when the routing information had been toggled on.

Looking at the code a property (`markerRefs`) was being set on the 
MapView to keep track of the map Markers. This was being destroyed 
for some reason when the routing toggle was pressed and since this
Array of references was being used to retrieve the pressed Marker so 
that it could be brought to the front this was leading to the crash.

Changed the way the refs are stored in the map so that they persist and
and made sure that when the map refreshes that the selected Marker’s 
`z-index` value is set higher than those that aren’t selected.