Restore video playback for android
Using WebView for playing video on Android causes the video to simply
play as in the thumbnail rather than full screen (like iOS. 

Previously this was worked around by creating an image thumbnail for the
video, overlaying another image with a play button on it and then 
playing the video in the image modal. This was broken because it was 
embedding an <Image> element within another <Image> element. 

By replacing the outer <Image> element with a <ImageBackground> element
we can work around this.